+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Control volume placement and accessibility by using CSIs Topology feature. Controlled by setting
+  [`csi.enableTopology`](./doc/helm-values.adoc#csienabletopology).
 * All pods use a dedicated service account to allow for fine-grained permission control.
 * The new [helm section `psp.*`](./doc/helm-values.adoc#psp) can automatically configure the ServiceAccount
   of all components to use the appropriate [PSP](https://kubernetes.io/docs/concepts/policy/pod-security-policy/) roles.